move authApi to API folder
